requirements.txt
amqp==5.0.6
anyjson==0.3.3
argcomplete==1.11.1
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.82
botocore==1.20.82
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-braces==1.14.0
django-celery-results==2.0.1
django-celery-beat==2.2.0
django-cors-headers==3.2.1
django-crispy-forms==1.10.0
django-db-logger==0.1.9
django-debug-toolbar==3.1.1
django-extensions==3.1.1
django-filter==2.4.0
django-jazzmin==2.4.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-s3-storage==0.13.2
django-simple-history==3.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.1
durationpy==0.5
factory-boy==3.2.0
Faker==6.6.0
future==0.18.2
gunicorn==19.9.0
gevent==20.9.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.4.0
google-cloud-tasks==2.0.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
hjson==3.0.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
install==1.3.3
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
kombu==5.0.2
langdetect==1.0.8
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pip-tools==5.2.1
placebo==0.9.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil
python-dotenv==0.17.0 
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
raven==6.10.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rollbar==0.15.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2
sqlparse==0.3.1
svg.path==4.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.47.0
troposphere==2.6.1
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.1.0
